requirements.txt
pillow==8.0.1
pickle==4.0

